Task: MNG-88 -- Integrate CG5251 Light Sensor - Kernel Bringup
Current status:
The driver is now able to interract with the sensor without causing a segmentation fault, thanks to the addition of the proper DT linkages and a correction of the conditional that tests i2c_check_functionality.

Current issue:
When probing the sensor for Lux data via sysfs or the input event interface, the device never reports the proper lux value. The sensor will always report 0 Lux no matter how bright the room is.

This commit is made up of the following work performed on the CG5251 Driver:
MNG-88 -- First attempt to bring up the CG5251 Light Sensor
MNG-88 -- Undefine CG_DBG in cg5251_poll.h in order to correct the sysfs permissions to avoid an assertion failure
MNG-88 -- Attempt to add proper device tree linkages to allow the driver to recognize the CG5251 device
MNG-88 -- cg5251_poll: i2c_check_functionality returns 1 on success, not zero. From i2c.h -- 'Return 1 if adapter supports everything we need, 0 if not.'